Added option of constructor usage for point creation

+++ OOP/Lab_3/Point.cpp
@@ -3,7 +3,7 @@ Point::Point(){
   x=0;
   y=0;
 }
-Point::Point(int a, int b){
+Point::Point(float a, float b){
     x=a;
     y=b;
 }
